requirements.txt
dash==2.9.2
numpy==1.24.3
plotly==5.14.0
pandas==2.0.0
gunicorn==20.1.0
dash_bootstrap_components==1.4.1
dash_leaflet==0.1.23
geopandas==0.13.0
dash_extensions==1.0.0
owslib==0.29.1
rasterio==1.3.6
pytorch==2.0.0
torchvision==0.15.1
tifffile==1.24.3
matplotlib==3.7.1
pandas==2.0.0

##python3.10 -m pip install GDAL==$(gdal-config --version)